requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==1.1.0
adafruit-circuitpython-motor==3.4.13
adafruit-circuitpython-motorkit==1.6.14
adafruit-circuitpython-pca9685==3.4.15
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==3.2.6
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
aniso8601==1.2.0
asn1crypto==1.5.1
blinker==1.8.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==1.1.1
Flask-Ask==0.8.8
idna==3.7
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.5
pycparser==2.22
pyftdi==0.55.4
pyOpenSSL==24.1.0
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
PyYAML==3.12
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.11.0
sysv-ipc==1.1.0
typing_extensions==4.11.0
Werkzeug==0.16.1
